+++ Flow.Launcher.Plugin.ShortcutPlugin/Models/Shortcuts/ShellShortcut.cs
@@ -1,26 +1,34 @@
-﻿namespace Flow.Launcher.Plugin.ShortcutPlugin.Models.Shortcuts;
+﻿using System.Text.Json.Serialization;
+
+namespace Flow.Launcher.Plugin.ShortcutPlugin.Models.Shortcuts;
+
+public enum ShellType
+{
+    Cmd,
+    Powershell
+}
 
 public class ShellShortcut : Shortcut
 {
-    public string Command { get; set; }
+    [JsonConverter(typeof(JsonStringEnumConverter))]
+    public ShellType ShellType { get; set; }
 
-    public string TargetFilePath { get; set; }
-
-    public bool Silent { get; set; }
+    public string Arguments { get; set; }
+    public bool Silent { get; set; } = true;
 
     public override object Clone()
     {
         return new ShellShortcut
         {
             Key = Key,
-            Command = Command,
-            TargetFilePath = TargetFilePath,
+            Arguments = Arguments,
+            ShellType = ShellType,
             Silent = Silent
         };
     }
 
     public override string ToString()
     {
-        return Command;
+        return $"{Arguments} ({ShellType.ToString().ToLower()})";
     }
 }