+++ Flow.Launcher.Plugin.ShortcutPlugin/Services/Interfaces/ICommandsService.cs
@@ -5,7 +5,7 @@ namespace Flow.Launcher.Plugin.ShortcutPlugin.Services.Interfaces;
 
 public interface ICommandsService
 {
-    List<Result> ResolveCommand([NotNull] List<string> arguments, string query);
+    List<Result> ResolveCommand([NotNull] List<string> arguments, Query query);
 
     void ReloadPluginData();
 }