+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Add quotes escaping
+
+## [1.1.3] - 2024-01-31
+
+- New shortcut type: `shell'. This type allows you to execute commands in command prompt or powershell.
+- Improved auto completion for commands.
+
 ## [1.1.2] - 2024-01-22
 
 - Fix group shortcut invocation